GREEN change parsing of individual instructions.
Instead of picking out characters one by one, we can
use the String.split method to achieve the same end.

This is a place where I ended up using the debugger;
my initial attempt used the wrong argument for
String.split, so I went in to see what the data looked
like.

It might have been "better" to stop and write a new
test to explore the behavior of that method, using
the output of the assert statement instead of the
debugging window.